pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "escalation-engine"
version = "1.0.0"
description = "Intelligent decision routing for 40x cost reduction through smart LLM usage"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Casey", email = "casey@users.noreply.github.com"}
]
maintainers = [
    {name = "Casey", email = "casey@users.noreply.github.com"}
]
keywords = [
    "llm",
    "routing",
    "escalation",
    "cost-optimization",
    "ai",
    "decision-engine",
    "fastapi",
    "multi-agent"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.9"
dependencies = [
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "httpx>=0.24.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.7.0",
    "ruff>=0.0.285",
    "mypy>=1.5.0",
    "pre-commit>=3.3.0",
]
openai = ["openai>=1.0.0"]
anthropic = ["anthropic>=0.18.0"]
ollama = ["ollama>=0.1.0"]
all = [
    "escalation-engine[openai,anthropic,ollama]",
]

[project.urls]
Homepage = "https://github.com/casey/escalation-engine"
Documentation = "https://github.com/casey/escalation-engine#readme"
Repository = "https://github.com/casey/escalation-engine.git"
Issues = "https://github.com/casey/escalation-engine/issues"

[project.scripts]
escalation-server = "escalation_engine.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["escalation_engine*"]
exclude = ["tests*", "examples*"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py39"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.isort]
known-first-party = ["escalation_engine"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore-missing-imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
